Unit testing of new feature extension approach
