+++ service-front/app/test/CommonTest/View/Twig/FeatureFlagExtensionTest.php
@@ -4,11 +4,10 @@ declare(strict_types=1);
 
 namespace CommonTest\View\Twig;
 
+use Common\View\Twig\FeatureFlagExtension;
+use Common\View\Twig\FeatureFlagRuntime;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
-use PHPUnit\Framework\Attributes\DataProvider;
-use Common\Service\Features\FeatureEnabled;
-use Common\View\Twig\FeatureFlagExtension;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Twig\TwigFunction;
@@ -21,16 +20,14 @@ class FeatureFlagExtensionTest extends TestCase
     #[Test]
     public function it_returns_an_array_of_exported_twig_functions(): void
     {
-        $extension = new FeatureFlagExtension(
-            $this->prophesize(FeatureEnabled::class)->reveal()
-        );
+        $extension = new FeatureFlagExtension();
 
         $functions = $extension->getFunctions();
 
         $this->assertIsArray($functions);
 
         $expectedFunctions = [
-            'feature_enabled' => 'featureEnabled',
+            'feature_enabled' => FeatureFlagRuntime::class . '::featureEnabled',
         ];
         $this->assertEquals(count($expectedFunctions), count($functions));
 
@@ -40,43 +37,7 @@ class FeatureFlagExtensionTest extends TestCase
             $this->assertArrayHasKey($function->getName(), $expectedFunctions);
 
             $functionCallable = $function->getCallable();
-            $this->assertInstanceOf(FeatureFlagExtension::class, $functionCallable[0]);
-            $this->assertEquals($expectedFunctions[$function->getName()], $functionCallable[1]);
+            $this->assertEquals($expectedFunctions[$function->getName()], $functionCallable);
         }
-    }
-
-    #[DataProvider('configuredFeatures')]
-    #[Test]
-    public function it_returns_the_features_configured_status(string $featureName, bool $enabled): void
-    {
-        $service = $this->prophesize(FeatureEnabled::class);
-        $service
-            ->__invoke($featureName)
-            ->willReturn($enabled);
-
-        $extension = new FeatureFlagExtension($service->reveal());
-
-        $result = $extension->featureEnabled($featureName);
-
-        $this->assertEquals($enabled, $result);
-    }
-
-    /**
-     * @return array<array<mixed>>
-     */
-    public static function configuredFeatures(): array
-    {
-        return [
-            'feature enabled'
-                => [
-                    'test_feature',
-                    false,
-                ],
-            'feature disabled'
-                => [
-                    'test_feature',
-                    true,
-                ],
-        ];
     }
 }
